python-script/requirement.txt
firebase-admin
google-cloud-firestore
icalendar
pytz
python-dateutil
